[psxtools] clut2png, rusty 2h, scaled dx,dy output

+++ tools/psxtools/clut.sh
@@ -1,15 +1,29 @@
 #!/bin/bash
-php="/tmp/clut2bmp.php"
-[ -f "$php" ] || exit
-export php
+php=""
+png=''
+if [ -f "/tmp/clut2png.php" ]; then
+	php="/tmp/clut2png.php"
+	png=1
+elif [ -f "/tmp/clut2bmp.php" ]; then
+	php="/tmp/clut2bmp.php"
+	png=''
+fi
+[ "$php" ] || exit
+export php png
 
 function clut2png
 {
 	[ -f "$1" ] || continue
-	php.sh $php "$1"
-	mogrify -format png -strip "$1".bmp &> /dev/null
+	php.sh  "$php"  "$1"
+
+	if [ ! "$png" ]; then
+		convert "$1.bmp" \
+			-define png:include-chunk=none,trns -strip \
+			"$1".png &> /dev/null
+		rm "$1".bmp
+	fi
+
 	rm "$1"
-	rm "$1".bmp
 }
 export -f clut2png
